Handle oscillation and oscillation_state differently on V2
Thanks to @Scaredycrow for finding this issue. It appears that V2
devices emit two states: oscillation (the configured mode; this is
the intent) and oscillation_state (what the fan is currently doing).
In most cases, these should be equivalent -- though the fan will
report "IDLE" if the fan is intended to oscillate but has met its
air quality/heat target & is thusly turned off.

To make this work: dyson_oscillation_mode now reflects the intended
mode (which is no change for V1 and not a significant change for V2),
and a new metric: dyson_oscillation_state covers the possibility of
IDLE. I also added code for V1 to simulate the V2 behaviour for
consistency between the model generations.