Switch to libdyson-based discovery
This is the second part of a three-part change to refactor this code
from libpurecool to libdyson. This part replaces the discovery components
(e.g; DysonDevice.auto_connect in libpurecool) to the separate API provided
by libdyson. This also adapts the libpurecool_adapter shim to have a
libdyson-like get_device() method to make switching fully over easier in
the next change.